+++ src/biz/bokhorst/xprivacy/PrivacyService.java
@@ -1342,6 +1342,14 @@ public class PrivacyService {
 					// Get application info
 					final ApplicationInfoEx appInfo = new ApplicationInfoEx(context, restriction.uid);
 
+					// Check for system application
+					if (appInfo.isSystem()) {
+						Version sVersion = new Version(getSetting(new PSetting(userId, "",
+								PrivacyManager.cSettingVersion, "0.0")).value);
+						if (sVersion.compareTo(new Version("2.0.38")) < 0)
+							return false;
+					}
+
 					// Check if activity manager agrees
 					if (!XActivityManagerService.canOnDemand())
 						return false;